requirements.txt
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
Flask==2.0.3
Flask-Caching==1.10.1
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
redis==4.2.0
requests==2.27.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
